% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eP-config-helper.R
\name{eP.config.helper}
\alias{eP.config.helper}
\title{Manages configuration of evalPurency with labpresets}
\usage{
eP.config.helper(
  labpreset,
  blankKey,
  sep,
  dec,
  colPol,
  colL,
  colReqPol,
  ReqPolKey,
  colShape,
  colCol,
  colLFib,
  colArea,
  colWidth,
  fibre,
  sphere,
  fragment,
  pixel,
  startrow
)
}
\arguments{
\item{labpreset}{A preset for most of the parameters (except: path, polymers, dataReturn, eocsum).
Can be requested by other labs, to be implemented, that they don't have to be set manually all the time.}

\item{blankKey}{The key word to distinguish blanks from other measurements/samples. It is case sensitive to prevent
accidental matching. Defaults to "Blank".}

\item{sep}{Symbol in your csv files indicating new a column. Defaults to ';'.}

\item{dec}{Symbol in your csv files indicating decimal sign. Defaults to ','.}

\item{colPol}{Column number where the polymer type is stated. In the TOEKI lab this is column 6 (Class Name).
The polymer names (or abbreviations) can be set manually in 'polymers' in case the default does not apply.
Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d).}

\item{colL}{Column number for the particle length. In the TOEKI lab this is column 17 (Length 5µ).
Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d).}

\item{colReqPol}{Column number for the particle check, whether the particle is a polymer or not.
In the TOEKI lab this is column 24 (Plastik? or Plastik ja/nein). Could also be provided as column name, but only
in ASCII encoding (e.g., special character as . and ä = d).}

\item{ReqPolKey}{key word or sentence of 'colReqPol' that indicates that it is a plastic particle. Default is 'ja'.}

\item{colShape}{Column number for the particle shape. In the TOEKI lab this is column 25 (Form).
Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d).}

\item{colCol}{Column number for the particle color In the TOEKI lab this is column 26 (Farbe).
Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d).}

\item{colLFib}{Column number for the particle length in case of a fibre with corrected length (because of curvy shape)
In the TOEKI lab this is column 27 (Länge). Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d).}

\item{colArea}{Column number for the particle length in case of a fibre with corrected length (because of curvy shape)
In the TOEKI lab this is column 4 (Area µm²). Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d).}

\item{colWidth}{Column number for the particle length in case of a fibre with corrected length (because of curvy shape)
In the TOEKI lab this is column 18 (Width µm). Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d).}

\item{fibre}{How fibres are called in colShape (Form). In the TOEKI lab it is 'Faser'.}

\item{sphere}{How spheres are called in colShape (Form). In the TOEKI lab it is 'Kugel'.}

\item{fragment}{How fragments are called in colShape (Form). In the TOEKI lab it is 'Fragment'.}

\item{pixel}{How pixels are called in colShape (Form). In the TOEKI lab it is 'Pixel'.}

\item{startrow}{The row where the data begin within the .csv file of purency (usually at 40).}
}
\description{
Most settings of evalPurency can be overwritten by a labpreset. If a labpreset is
selected, it might overwrite or not overwrite the current function settings.
To manage which settings will be chosen and which not (depending on NAs in labpreset),
this function handles the configuration with labpresets correctly and returns a config data frame.
This function is thought to ensure correct configuration and to simplify changes and maintenance,
as well as improving testing.
}
