% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalPurency.R
\name{evalPurency}
\alias{evalPurency}
\title{created by Marvin Kiene, use this function at own risk!}
\usage{
evalPurency(
  path,
  polymers = c("PU", "EVAc", "PA", "PAN", "PBT", "PET", "PE", "PMMA", "PP", "POM", "PS",
    "PVC", "PC", "ABS", "PPSU", "CA", "PEEK", "EVOH", "PSU", "SI", "PLA", "PLAPBAT"),
  dataReturn = FALSE
)
}
\arguments{
\item{path}{The path where the files can be found. All csv files in this folder will be evaluated. Also saves the
resulting files in this directory.}

\item{polymers}{A vector containing the abbreviations of polymers to be considered. Default vector contains 22
polymers.}

\item{dataReturn}{If set TRUE, a data frame will be returned containing the data of all measurement with
the necessary information.}
}
\description{
Evaluate csv files, produced by purrency. It will count occurences of of fibres, fragments, spheres and pixels,
as well as size fractions (<10, 10-20, 20-50, 50-100, 100-150, 150-200,..., >500) for each polymer. Each file (i.e., each measurement)
is evaluated separately, as well as summarized for all files (i.e., one sample).
}
