% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eP-load-helper.R
\name{ep.load.helper}
\alias{ep.load.helper}
\title{Manages data loading and pre-processing of the evalPurency package. In the end, also a quality control is
performed that throws warnings if there are empty files or fields, including the name and column.}
\usage{
ep.load.helper(
  path,
  particleNumbers,
  sep,
  dec,
  formFillDefault,
  colourFillDefault,
  colL,
  colPol,
  startrow,
  colReqPol,
  colourSep,
  ReqPolKey,
  colShape,
  colCol,
  colLFib,
  colArea,
  colWidth,
  test,
  fibre,
  sphere,
  fragment,
  pixel
)
}
\arguments{
\item{path}{The path set by the user in main function.}

\item{particleNumbers}{set TRUE if you would like to get an extra file with just plastic and non-plastic
particle numbers for each file loaded.}

\item{sep}{Column separator within the *.csv files. Defaults to ';' in main function.}

\item{dec}{Decimal sign within the *.csv files. Defaults to ',' in main function.}

\item{formFillDefault}{If desired you can provide a default form that will be filled in, if no form is provided (NA). Should be
one of the values of the parameters 'fibre', 'sphere', 'fragment' or 'pixel' (see below).}

\item{colourFillDefault}{If desired you can provide a default colour that will be filled in, if no colour is provided in the data (NA).
However, in this function colour has no further meansing.}

\item{colL}{Column number for the particle length. In the TOEKI lab this is column 17 (Length 5µ).
Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d).
Set in main function.}

\item{colPol}{Column number where the polymer type is stated. In the TOEKI lab this is column 6 (Class Name).
The polymer names (or abbreviations) can be set manually in 'polymers' in case the default does not apply.
Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d).
Set in main function.}

\item{startrow}{Number of rows that can be omitted from the *.csv files. Usually preprocessing software
provides a bunch of meta data, that are not of interest here. Usually 40 rows can be skipped (more or less - 0 in case of siMPle).
This is automated now. However, automation can be disabled if a value is defined (e.g., for troubleshooting).}

\item{colReqPol}{Column number for the particle check, whether the particle is a polymer or not.
In the TOEKI lab this is column 24 (Plastik? or Plastik ja/nein). Could also be provided as column name, but only
in ASCII encoding (e.g., special character as . and ä = d). Set in main function.}

\item{colourSep}{whether colour separation is desired for the analysis. Only usable in evalPurency.particles()!}

\item{ReqPolKey}{key word or sentence of 'colReqPol' that indicates that it is a plastic particle.
Default is 'ja'. Set in main function.}

\item{colShape}{Column number for the particle shape. In the TOEKI lab this is column 25 (Form).
Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d).
Set in main function.
Can also be set to 'none' if no column for shape exists in the files. This will cause the function to create one and set as
'fragment' as shape. However, this can be ignored afterwards.}

\item{colCol}{Column number for the particle color In the TOEKI lab this is column 26 (Farbe).
Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d).
Set in main function.}

\item{colLFib}{Column number for the particle length in case of a fibre with corrected length (because of curvy shape)
In the TOEKI lab this is column 27 (Länge). Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d).
Set in main function.}

\item{colArea}{Column number for the particle length in case of a fibre with corrected length (because of curvy shape)
In the TOEKI lab this is column 4 (Area µm²). Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d).
Set in main function.}

\item{colWidth}{Column number for the particle length in case of a fibre with corrected length (because of curvy shape)
In the TOEKI lab this is column 18 (Width µm). Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d).
Set in main function.}

\item{test}{Can be set TRUE when the function should be run in testing mode. Set in main function.}

\item{fibre}{How fibres are called in colShape (Form). In the TOEKI lab it is 'Faser'.}

\item{sphere}{How spheres are called in colShape (Form). In the TOEKI lab it is 'Kugel'.}

\item{fragment}{How fragments are called in colShape (Form). In the TOEKI lab it is 'Fragment'.}

\item{pixel}{How pixels are called in colShape (Form). In the TOEKI lab it is 'Pixel'.}
}
\value{
Returns a data frame containing a list of all particles and their properties.
}
\description{
Shared by both main function of the evalPurency package. Loads data from all *.csv files from the
provided folter.
}
